Fix sign warnings for e2.
Like d2s.c, normal ieeeExponent values are within [1, 2046], so we
should have a value-preserving cast from uint32_t to int32_t before
subtracting to produce possibly negative values.

+++ ryu/d2fixed.c
@@ -377,7 +377,7 @@ int d2fixed_buffered_n(double d, uint32_t precision, char* result) {
     e2 = 1 - DOUBLE_BIAS - DOUBLE_MANTISSA_BITS;
     m2 = ieeeMantissa;
   } else {
-    e2 = ieeeExponent - DOUBLE_BIAS - DOUBLE_MANTISSA_BITS;
+    e2 = (int32_t) ieeeExponent - DOUBLE_BIAS - DOUBLE_MANTISSA_BITS;
     m2 = (1ull << DOUBLE_MANTISSA_BITS) | ieeeMantissa;
   }
 
@@ -586,7 +586,7 @@ int d2exp_buffered_n(double d, uint32_t precision, char* result) {
     e2 = 1 - DOUBLE_BIAS - DOUBLE_MANTISSA_BITS;
     m2 = ieeeMantissa;
   } else {
-    e2 = ieeeExponent - DOUBLE_BIAS - DOUBLE_MANTISSA_BITS;
+    e2 = (int32_t) ieeeExponent - DOUBLE_BIAS - DOUBLE_MANTISSA_BITS;
     m2 = (1ull << DOUBLE_MANTISSA_BITS) | ieeeMantissa;
   }
